test_task/pyproject.toml
[project]
name = "test-task"
version = "0.1.0"
description = ""
authors = [
    {name = "IDaniil228",email = "99666220+IDaniil228@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (==4.1)",
    "djangorestframework (==3.15.1)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "psycopg (>=3.3.3,<4.0.0)",
    "django-pg-upsert (>=0.5.0,<0.6.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
